Remove chipmunk include
On macOS this brings Objective-C headers in the closure which fails to
compile with our C source file.

Another solution would be to pick a Objective-C version of the
cpfloat_limits.c for macOS. But since we always expect cpFloat to be
double (hardcoded in Ada code), this is good enough for now.